+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * allow to use `set-repr!` on records (record type name is now a class)
 * match function return `#f` instead of `nil` when fail to match (so it work in `cond` or `if`)
 * new functions `complement`, `always`, `once`, `flip`, `unfold` (inspired by Ramda)
+* experimental compiler to JSON (not 100% ready)
 ### Bugfix
 * fix `set-repr!` on classes
 * fix conflict with jQuery plugns in bookmark